move helper classes to its own file
