+++ src/sc_keeper/api.py
@@ -10,16 +10,7 @@ from typing import Annotated, List, Literal, Optional
 from fastapi import Depends, FastAPI, HTTPException, Path, Query, Request, Response
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.middleware.gzip import GZipMiddleware
-from pydantic import BaseModel
-from sc_crawler.table_bases import (
-    BenchmarkScoreBase,
-    CountryBase,
-    RegionBase,
-    ServerBase,
-    ServerPriceBase,
-    VendorBase,
-    ZoneBase,
-)
+from sc_crawler.table_bases import ServerBase
 from sc_crawler.table_fields import Allocation, CpuArchitecture, Status, StorageType
 from sc_crawler.tables import (
     Benchmark,
@@ -44,6 +35,19 @@ from .database import session
 from .logger import LogMiddleware, get_request_id
 from .lookups import min_server_price
 from .query import max_score_per_server
+from .references import (
+    Countries,
+    Vendors,
+    Regions,
+    ComplianceFrameworks,
+    ServerTableMetaData,
+    ServerPKs,
+    ServerPKsWithPrices,
+    RegionPKs,
+    ServerPriceWithPKs,
+    OrderDir,
+    FilterCategories,
+)
 
 package_versions = {
     pkg: version(pkg)
@@ -84,100 +88,6 @@ session.updated.wait()
 
 # ##############################################################################
 # Helper classes
-
-# create enums from DB values for filtering options
-Countries = StrEnum(
-    "Countries", {m.country_id: m.country_id for m in db.exec(select(Country)).all()}
-)
-Vendors = StrEnum(
-    "Vendors", {m.vendor_id: m.vendor_id for m in db.exec(select(Vendor)).all()}
-)
-Regions = StrEnum(
-    "Regions",
-    {m.region_id: m.region_id for m in db.exec(select(Region)).all()},
-)
-ComplianceFrameworks = StrEnum(
-    "ComplianceFrameworks",
-    {
-        m.compliance_framework_id: m.compliance_framework_id
-        for m in db.exec(select(ComplianceFramework)).all()
-    },
-)
-
-
-class NameAndDescription(BaseModel):
-    name: str
-    description: str
-
-
-class IdNameAndDescription(NameAndDescription):
-    id: str
-
-
-class TableMetaData(BaseModel):
-    table: NameAndDescription
-    fields: List[IdNameAndDescription]
-
-
-class IdNameAndDescriptionAndCategory(IdNameAndDescription):
-    category: str
-    unit: Optional[str]
-
-
-class ServerTableMetaData(TableMetaData):
-    fields: List[IdNameAndDescriptionAndCategory]
-
-
-class ServerWithScore(ServerBase):
-    score: Optional[float] = None
-    price: Optional[float] = None
-    score_per_price: Optional[float] = None
-
-
-class ServerPKs(ServerWithScore):
-    vendor: VendorBase
-
-
-class ServerPricePKs(ServerPriceBase):
-    region: RegionBase
-    zone: ZoneBase
-
-
-class ServerPKsWithPrices(ServerPKs):
-    prices: List[ServerPricePKs]
-    benchmark_scores: List[BenchmarkScoreBase]
-
-
-class RegionPKs(RegionBase):
-    vendor: VendorBase
-
-
-class RegionBaseWithPKs(RegionBase):
-    country: CountryBase
-
-
-class ServerPriceWithPKs(ServerPriceBase):
-    vendor: VendorBase
-    region: RegionBaseWithPKs
-    zone: ZoneBase
-    server: ServerWithScore
-
-
-class OrderDir(Enum):
-    ASC = "asc"
-    DESC = "desc"
-
-
-class FilterCategories(Enum):
-    BASIC = "basic"
-    PRICE = "price"
-    PROCESSOR = "processor"
-    MEMORY = "memory"
-    REGION = "region"
-    VENDOR = "vendor"
-    STORAGE = "storage"
-    GPU = "gpu"
-
 
 # load examples for the docs
 example_data = {
